pyproject.toml
[project]
name = "ballsdex"
description = ""
authors = [
    { name = 'Auguste "laggron42" Charpentier', email = "laggron42@ballsdex.com" },
    { name = 'Jamie "flare" McGuinness', email = "flare@ballsdex.com" },
]
license = "MIT"
requires-python = ">=3.13, <3.14"
dynamic = ["version"]
dependencies = [
    # asyncio
    "uvloop==0.21.0; sys_platform != 'win32'",

    # discord
    "discord.py[speed]==2.5.0",

    # database ORM
    "tortoise-orm[asyncpg] ==0.22.2",
    "tortoise-cli==0.1.2",

    # fuzzy matching
    "rapidfuzz==3.6.2",

    # django admin panel
    "uvicorn[standard]==0.32.0",
    "django==5.1.4",
    "django-allow-cidr (>=0.7.1,<0.8.0)",
    "django-nonrelated-inlines==0.2",
    "django-admin-autocomplete-filter==0.7.1",
    "django-admin-action-forms==1.3.0",
    "django-admin-inline-paginator==0.4.0",
    "dj-database-url==2.3.0",
    "social-auth-app-django==5.4.2",
    "psycopg==3.2.3",
    "psycopg-binary==3.2.3",

    # metrics
    "prometheus-client==0.20.0",

    # misc
    "rich==13.8.0",
    "python-dateutil==2.9.0",
    "Pillow==10.4.0",
    "pyyaml==6.0.2",
    "cachetools==5.5.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit==3.7.1",
    "black==24.8.0",
    "flake8-pyproject==1.2.3",
    "flake8==7.2.0",
    "pyright==1.1.390",
    "isort==5.13.2",
    "django-debug-toolbar==4.4.6",
    "pyinstrument==5.0.0",
    "django-types==0.20.0",
]

[tool.poetry]
packages = [{ include = "ballsdex" }]
requires-poetry = ">=2.0"
version = "0.0.0"

[tool.poetry.requires-plugins]
poetry-plugin-shell = ">=1.0"
poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }

[tool.poetry-dynamic-versioning]
enable = true

[tool.poetry-dynamic-versioning.from-file]
source = "ballsdex/__init__.py"
pattern = '__version__ = "(?P<base>\d+\.\d+\.\d+)"'

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99

[tool.flake8]
ignore = "W503,E203,E999"
max-line-length = 99
exclude = "./admin_panel/admin_panel/settings/*"

[tool.isort]
profile = "black"
line_length = 99

[tool.pyright]
extraPaths = ["./admin_panel"]
pythonVersion = "3.13"
reportIncompatibleMethodOverride = "warning"
reportIncompatibleVariableOverride = "warning"
